migration fixed

+++ database/migrations/2026_08_19_080502_create_whatsapp_inboxes_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        // Skip this migration if running on master database
+        if (Schema::getConnection()->getName() === 'mysql') {
+            return;
+        }
         Schema::create('whatsapp_inboxes', function (Blueprint $table) {
             $table->id();
             $table->string('sender_number');
@@ -30,6 +34,10 @@ return new class extends Migration
      */
     public function down(): void
     {
+        // Skip this migration if running on master database
+        if (Schema::getConnection()->getName() === 'mysql') {
+            return;
+        }
         Schema::dropIfExists('whatsapp_inboxes');
     }
 };
